Reorganizing how plugins are used. Previously a plugin was a specific check. Now a plugin will be what was referred to previously a "Plugin Type". That means for example, CheckAbapTableEntryCount will be the actual plugin.
The actual checks will be stored in the database. This has a lot of advantages. In particular the forms for generating or configuring a plugin can be automatically generated (as they are in the system settings).

It's a bit more complicated to implement, but will result in a lot more flexibility down the road,